requirements.txt
tensorboardX
tensorboard
scikit-learn
seqeval
wget
tqdm
ptvsd
torch==1.9.0
boto3
filelock
tokenizers
regex
streamlit
PyYAML
pydantic